V7.22.7: Abgleich SD_Card_Lib
Auto-commit: 3 file(s) in Horst

Timestamp: 2026-09-21 17:32

Changed files:
  [modified] Horst/Flash_F4xx_Lib.c
  [modified] Horst/Flash_F4xx_Lib.h
  [modified] Horst/Portable_Utils.c

+++ Horst/Portable_Utils.c
@@ -709,7 +709,7 @@ void portable_memset ( uint8_t * dest, uint8_t value, uint32_t size )
 //
 int16_t  portable_memcmp ( uint8_t * dest, uint8_t * source, uint32_t size )
 {
-  int16_t  i       = 0;
+  int32_t  i       = 0;
   int16_t  result  = 0;
 
   while ( i < size )
